requirements-dev.txt
pre-commit>=3.6.0
ruff>=0.3.0
mypy>=1.8.0
pytest>=8.0.0
pytest-cov>=4.1.0
tabulate>=0.9.0
pyyaml>=6.0
types-PyYAML>=6.0
types-tabulate>=0.9
